kernel: make all sizes unsigned, sort out the sign mess

+++ src/init/syscalls.c
@@ -22,12 +22,12 @@ int _syscall_mount(handle_t handle, const char __user *path, int len) {
 	return _syscall(_SYSCALL_MOUNT, handle, (int)path, len, 0);
 }
 
-int _syscall_read(handle_t handle, void __user *buf, int len, int offset) {
-	return _syscall(_SYSCALL_READ, handle, (int)buf, len, offset);
+int _syscall_read(handle_t handle, void __user *buf, size_t len, int offset) {
+	return _syscall(_SYSCALL_READ, handle, (int)buf, (int)len, offset);
 }
 
-int _syscall_write(handle_t handle, const void __user *buf, int len, int offset) {
-	return _syscall(_SYSCALL_WRITE, handle, (int)buf, len, offset);
+int _syscall_write(handle_t handle, const void __user *buf, size_t len, int offset) {
+	return _syscall(_SYSCALL_WRITE, handle, (int)buf, (int)len, offset);
 }
 
 int _syscall_close(handle_t handle) {
